add common datepicker components
