+++ src/components/DatePicker/Common.tsx
@@ -1,0 +1,62 @@
+import styled from "styled-components";
+import { InputElement, InputWrapper } from "../Input/InputWrapper";
+import { useId } from "react";
+import { Icon } from "../Icon/Icon";
+
+const locale = "en-US";
+const selectedDateFormatter = new Intl.DateTimeFormat(locale, {
+  day: "2-digit",
+  month: "short",
+  year: "numeric",
+});
+
+const explicitWidth = "250px";
+
+const HighlightedInputWrapper = styled(InputWrapper)<{ $isActive: boolean }>`
+  ${({ $isActive, theme }) => {
+    return `border: ${theme.click.datePicker.dateOption.stroke} solid ${
+      $isActive
+        ? theme.click.datePicker.dateOption.color.stroke.active
+        : theme.click.field.color.stroke.default
+    };`;
+  }}
+
+  width: ${explicitWidth};
+}`;
+
+interface DatePickerInputProps {
+  isActive: boolean;
+  disabled: boolean;
+  id?: string;
+  placeholder?: string;
+  selectedDate?: Date;
+}
+
+export const DatePickerInput = ({
+  isActive,
+  disabled,
+  id,
+  placeholder,
+  selectedDate,
+}: DatePickerInputProps) => {
+  const defaultId = useId();
+  const formattedSelectedDate =
+    selectedDate instanceof Date ? selectedDateFormatter.format(selectedDate) : "";
+
+  return (
+    <HighlightedInputWrapper
+      $isActive={isActive}
+      disabled={disabled}
+      id={id ?? defaultId}
+    >
+      <Icon name="calendar" />
+      <InputElement
+        data-testid="datepicker-input"
+        placeholder={placeholder}
+        readOnly
+        value={formattedSelectedDate}
+      />
+    </HighlightedInputWrapper>
+  );
+};
+
